+++ Mods ChoGGi/Add 5 Times The Resource To Deposits/metadata.lua
@@ -8,19 +8,18 @@ return PlaceObj("ModDef", {
 		}),
 	},
 	"title", "Add 5 Times The Resource To Deposits",
+	"id", "ChoGGi_Add5TimesTheResourceToDeposits",
+	"pops_any_uuid", "6ff9709d-3a3b-48b5-be2f-ce3216aa5699",
+	"steam_id", "1427609324",
+	"lua_revision", 1001514, -- Tito
 	"version", 6,
 	"version_major", 0,
 	"version_minor", 6,
-
-	"id", "ChoGGi_Add5TimesTheResourceToDeposits",
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua"
 	},
 	"image", "Preview.png",
-	"steam_id", "1427609324",
-	"pops_any_uuid", "6ff9709d-3a3b-48b5-be2f-ce3216aa5699",
-	"lua_revision", 1001514, -- Tito
 	"description", [[Adds a button to water/metal/concrete deposits to multiple the amount of all deposits of the same resource by 5.
 
 Requested by Peacemaker.]],