+++ Mods ChoGGi/Autonomous Drones/items.lua
@@ -6,7 +6,7 @@ local concat = table.concat
 return {
 	PlaceObj("ModItemOptionToggle", {
 		"name", "EnableMod",
-		"DisplayName", T(302535920011303, "Enable Mod"),
+		"DisplayName", T(302535920011303, "<color ChoGGi_yellow>Enable Mod</color>"),
 		"Help", T(302535920011793, "Disable mod without having to see missing mod msg."),
 		"DefaultValue", true,
 	}),