+++ Mods ChoGGi/Block Pets/metadata.lua
@@ -4,10 +4,10 @@ return PlaceObj("ModDef", {
 	"steam_id", "1790982885",
 	"pops_any_uuid", "80075020-7955-46c2-9392-60a3460871a9",
 	"lua_revision", 1001514, -- Tito
-	"version", 2,
+	"version", 3,
 	"version_major", 0,
-	"version_minor", 2,
-	"image", "Preview.png",
+	"version_minor", 3,
+	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
@@ -15,6 +15,7 @@ return PlaceObj("ModDef", {
 	"has_options", true,
 	"TagOther", true,
 	"description", [[Pets will no longer spawn from service/residence buildings.
+Mod options to choose allowed pets. Less pets will spawn, but there's also a mod option to increase spawn rate.
 
 
 Requested by paladinBoyd.]],