+++ Mods ChoGGi/Fix Canceled Rocket On Pad/items.lua
@@ -1,7 +1,7 @@
 return {
 	PlaceObj("ModItemOptionToggle", {
 		"name", "EnableMod",
-		"DisplayName", T(302535920011303, "Enable Mod"),
+		"DisplayName", T(302535920011303, "<color ChoGGi_yellow>Enable Mod</color>"),
 		"Help", T(302535920011793, "Disable mod without having to see missing mod msg."),
 		"DefaultValue", true,
 	}),