+++ Mods ChoGGi/Nursery Limit Birthing To Spots/Code/Script.lua
@@ -1,20 +1,15 @@
 -- See LICENSE for terms
 
-local table = table
 local Max = Max
+local MapGet = MapGet
 
 local mod_GlobalDomeCount
-local mod_RespectIncubator
 local mod_BypassNoNurseries
-local mod_UltimateNursery
 
 -- fired when settings are changed/init
 local function ModOptions()
-	local options = CurrentModOptions
-	mod_GlobalDomeCount = options:GetProperty("GlobalDomeCount")
-	mod_RespectIncubator = options:GetProperty("RespectIncubator")
-	mod_BypassNoNurseries = options:GetProperty("BypassNoNurseries")
-	mod_UltimateNursery = options:GetProperty("UltimateNursery")
+	mod_GlobalDomeCount = CurrentModOptions:GetProperty("GlobalDomeCount")
+	mod_BypassNoNurseries = CurrentModOptions:GetProperty("BypassNoNurseries")
 end
 
 -- load default/saved settings
@@ -32,36 +27,28 @@ end
 local orig_SpawnChild = Dome.SpawnChild
 function Dome:SpawnChild(...)
 	-- hi Ski
-	if mod_RespectIncubator and self.IncubatorReloc then
+	if self.IncubatorReloc then
 		return orig_SpawnChild(self, ...)
-	end
-	local city = self.city or UICity
-
-	local class = mod_UltimateNursery and "UltimateNursery" or "Nursery"
-
-	local objs_g = table.icopy(city.labels[class] or empty_table)
-	if not mod_UltimateNursery then
-		-- why make a label with all the nursery varients, when you can cause extra work for modders...
-		table.append(objs_g, city.labels.LargeNursery or empty_table)
 	end
 
 	local objs
 
 	-- check all nurseries or per-dome
 	if mod_GlobalDomeCount then
-		objs = objs_g
+		objs = MapGet("map", "Nursery")
 	else
-		objs = table.icopy(self.labels[class] or empty_table)
-		if not mod_UltimateNursery then
-			table.append(objs, self.labels.LargeNursery or empty_table)
-		end
+		objs = MapGet("map", "Nursery", function(obj)
+			if obj.parent_dome == self then
+				return true
+			end
+		end)
 	end
 
 	local objs_count = #objs
 
 	-- no nursery so abort
 	if objs_count == 0 then
-		-- If there's no nurseries at all and then send back orig func
+		-- If there's no nurseries then send back orig func so births still work
 		if mod_BypassNoNurseries then
 			return orig_SpawnChild(self, ...)
 		end
@@ -79,8 +66,6 @@ function Dome:SpawnChild(...)
 		end
 	end
 	local free_space = Max(0, capacity - used)
-
---~ 	print(free_space,ChoGGi.ComFuncs.RetName(self))
 
 	-- fire up the babby factory
 	if free_space > 0 then