+++ Mods ChoGGi/Nursery Limit Birthing To Spots/items.lua
@@ -8,19 +8,7 @@ return {
 	PlaceObj("ModItemOptionToggle", {
 		"name", "BypassNoNurseries",
 		"DisplayName", T(302535920011454, "Bypass No Nurseries"),
-		"Help", T(302535920011510, "If there's no nurseries act like vanilla."),
+		"Help", T(302535920011510, "If there's no nurseries act like vanilla (otherwise no nursery == no births)."),
 		"DefaultValue", true,
-	}),
-	PlaceObj("ModItemOptionToggle", {
-		"name", "RespectIncubator",
-		"DisplayName", T(302535920011455, "Respect Incubator"),
-		"Help", T(302535920011511, "Allow SkiRich's Incubator mod to relocate newborns to incubator dome."),
-		"DefaultValue", false,
-	}),
-	PlaceObj("ModItemOptionToggle", {
-		"name", "UltimateNursery",
-		"DisplayName", T(302535920011456, "Ultimate Nursery"),
-		"Help", T(302535920011512, "Count Ultimate Nurseries INSTEAD of regular ones."),
-		"DefaultValue", false,
 	}),
 }