+++ Mods ChoGGi/Remove Landscaping Limit/items.lua
@@ -9,7 +9,7 @@ return {
 	}),
 	PlaceObj("ModItemOptionToggle", {
 		"name", "RemoveLandScapingLimits",
-		"DisplayName", T(302535920011303, "Enable Mod"),
+		"DisplayName", T(302535920011303, "<color ChoGGi_yellow>Enable Mod</color>"),
 		"Help", T(302535920011793, "Disable mod without having to see missing mod msg."),
 		"DefaultValue", true,
 	}),