+++ Mods ChoGGi/Skip Founder Stage/metadata.lua
@@ -1,0 +1,20 @@
+return PlaceObj("ModDef", {
+	"title", "Skip Founder Stage",
+	"id", "ChoGGi_SkipFounderStage",
+	"steam_id", "2529641398",
+	"pops_any_uuid", "d09154a1-db64-49eb-b5e5-f29f47d02d22",
+	"lua_revision", 1001514, -- Tito
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+Colony approval happens as soon as you load your save (or start a new game).
+]],
+})