+++ Expanded Cheat Menu/Changelog.md
@@ -1,6 +1,11 @@
 ## ECM Changelog
 
 ## v17.1 (Unreleased)
+### Changed:
+ - ConsoleSkipUndefinedGlobals defaults to false.
+
+### Fixed:
+- Fill resource didn't work on deposits (thanks Dhamptastic).
 
 ## v17.0 (21 June 2021)
 ### Changed: